Some reorganizing for adding the AllTermsTab
