+++ frontend/packages/portal-frontend/src/geneTea/components/GeneTea.tsx
@@ -12,6 +12,7 @@ import { fetchMetadata } from "../utils";
 import glossary from "src/geneTea/json/glossary.json";
 import Glossary from "src/common/components/Glossary";
 import { GlossaryItem } from "src/common/components/Glossary/types";
+import { TopTermsContextProvider } from "../context/TopTermsContext";
 
 function GeneTea() {
   const { handleSetAllAvailableGenes } = useGeneTeaFiltersContext();
@@ -72,9 +73,7 @@ function GeneTea() {
               </TabList>
               <TabPanels className={styles.TabPanels}>
                 <TabPanel className={styles.TabPanel}>
-                  <TopTermsContextProvider>
-                    <GeneTeaMainContent tab="top-tea-terms" />
-                  </TopTermsContextProvider>
+                  <GeneTeaMainContent tab="top-tea-terms" />
                 </TabPanel>
                 <TabPanel className={styles.TabPanel}>
                   <GeneTeaMainContent tab="all-matching-terms" />
